requirements.txt
lancedb
weaviate-client
streamlit
chainlit
bs4
langchain
google-search-results
openai
python-dotenv
lxml
html2text
tiktoken
gdown
google-api-python-client
#accelerate
